move positions page to committee layout
